+++ raspberrypi/install-skies-adsb.sh
@@ -4,12 +4,53 @@
 # this is the skies-adsb install script for the Raspberry Pi
 #
 
-if ! grep -q "Raspberry Pi" /proc/cpuinfo; then
-  echo "This script must be run on a Raspberry Pi"
+# if ! grep -q "Raspberry Pi" /proc/cpuinfo; then
+#   echo "This script must be run on a Raspberry Pi"
+#   exit 1
+# fi
+
+# Process command line options
+while getopts ":srde:" opt; do
+  case $opt in
+  s)
+    SKIP_RPI_UPGRADE=1
+    ;;
+  r)
+    ADSB_DRIVER="readsb"
+    ;;
+  d)
+    ADSB_DRIVER="dump1090"
+    ;;
+  e)
+    ADSB_DRIVER="existing"
+    ADSB_HOST_PORT="$OPTARG"
+    # Validate IP:port format
+    if ! echo "$ADSB_HOST_PORT" | grep -qE '^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+:[0-9]+$'; then
+      echo "Error: -e requires valid IP:port format (e.g. 192.168.1.123:30003)"
+      exit 1
+    fi
+    ;;
+  \?)
+    echo "Invalid option: -$OPTARG"
+    exit 1
+    ;;
+  :)
+    echo "Option -$OPTARG requires an argument"
+    exit 1
+    ;;
+  esac
+done
+
+if [ -z "$ADSB_DRIVER" ]; then
+  echo "Error: ADSB driver not specified. Use -r for readsb or -d for dump1090 or -e for using existing ADS-B receiver"
   exit 1
 fi
 
-optional_do_upgrade_rpi() {
+if [ -z "$ADSB_HOST_PORT" ]; then
+  ADSB_HOST_PORT="0.0.0.0:30003"
+fi
+
+upgrade_rpi() {
   echo "###############################################"
   echo "Updating and upgrading Raspberry Pi system..."
   echo "-----------------------------------------------"
@@ -27,7 +68,23 @@ optional_do_upgrade_rpi() {
   echo "**********************************************"
 }
 
-optional_install_dump1090() {
+install_readsb() {
+  echo "###############################################"
+  echo "Installing readsb..."
+  echo "-----------------------------------------------"
+
+  sudo bash -c "$(wget -O - https://github.com/wiedehopf/adsb-scripts/raw/master/readsb-install.sh)"
+
+  source ~/skies-adsb/src/.env
+
+  sudo readsb-set-location $VITE_DEFAULT_ORIGIN_LATITUDE $VITE_DEFAULT_ORIGIN_LONGITUDE
+
+  echo
+  echo "readsb installation complete!"
+  echo "**********************************************"
+}
+
+install_dump1090() {
   echo "###############################################"
   echo "Installing dump1090-mutability..."
   echo "-----------------------------------------------"
@@ -43,7 +100,7 @@ optional_install_dump1090() {
   echo "**********************************************"
 }
 
-do_setup_python_environment() {
+setup_python_environment() {
   echo "###############################################"
   echo "Setup Python environment..."
   echo "-----------------------------------------------"
@@ -69,7 +126,7 @@ do_setup_python_environment() {
   echo "**********************************************"
 }
 
-do_setup_app() {
+setup_app_start() {
   echo "###############################################"
   echo "Setting up skies-adsb..."
 
@@ -77,16 +134,16 @@ do_setup_app() {
   cd
   rm -rf skies-adsb
   mkdir -p skies-adsb
-  mv skies-*.{sh,service,tar.gz} skies-adsb/
+
+  echo "Extracting skies-adsb app..."
+  tar zxvf skies-adsb-app.tar.gz -C skies-adsb
+  rm skies-adsb-app.tar.gz
 
   # Setup Python environment
-  do_setup_python_environment
+  setup_python_environment
+}
 
-  # Extract and setup Flask application
-  echo "Setting up skies-adsb flask app..."
-  cd ~/skies-adsb
-  tar zxvf skies-adsb-flask-app.tar.gz
-  rm skies-adsb-flask-app.tar.gz
+setup_app_finish() {
   cd
 
   # Clean up existing services
@@ -95,6 +152,9 @@ do_setup_app() {
     sudo systemctl stop skies-adsb-${service}
     sudo rm -f /etc/systemd/system/skies-adsb-${service}.service
   done
+
+  echo "Replacing ADSB_HOST_PORT in websockify service script..."
+  sed -i "s/#ADSB_HOST_PORT#/${ADSB_HOST_PORT}/g" skies-adsb/skies-adsb-websockify.sh
 
   # Setup new system services
   echo "Setting up skies-adsb system services..."
@@ -109,7 +169,7 @@ do_setup_app() {
 
   echo "**********************************************"
   echo "Cleaning up and rebooting Raspberry Pi to complete setup..."
-  rm install.sh
+  rm install-skies-adsb.sh
   sudo reboot
 }
 
@@ -122,11 +182,30 @@ do_setup_app() {
 echo "Starting skies-adsb installation..."
 echo "===================================="
 
-optional_do_upgrade_rpi
+if [ -z "$SKIP_RPI_UPGRADE" ]; then
+  upgrade_rpi
+  echo
+fi
+
+setup_app_start
 echo
 
-optional_install_dump1090
+case "$ADSB_DRIVER" in
+"readsb")
+  install_readsb
+  ;;
+"dump1090")
+  install_dump1090
+  ;;
+"existing")
+  echo "###############################################"
+  echo "Using existing ADS-B receiver..."
+  echo "**********************************************"
+  ;;
+esac
 echo
 
-do_setup_app
+setup_app_finish
+echo
+
 echo "Installation complete!"