+++ raspberrypi/skies-adsb-websockify.sh
@@ -3,11 +3,8 @@
 # Websocket proxy for ADS-B SBS data
 #
 
-# Define constants
-LISTEN_HOST="0.0.0.0"
-LISTEN_PORT=30006
-TARGET_HOST="0.0.0.0"
-TARGET_PORT=30003
+LISTEN_HOST_PORT="0.0.0.0:30006"
+ADSB_HOST_PORT="#ADSB_HOST_PORT#"
 
 # Check if websockify is installed
 if ! command -v websockify >/dev/null 2>&1; then
@@ -16,5 +13,5 @@ if ! command -v websockify >/dev/null 2>&1; then
 fi
 
 # Start websockify
-echo "Starting websocket proxy on ${LISTEN_HOST}:${LISTEN_PORT}"
-websockify "${LISTEN_HOST}:${LISTEN_PORT}" "${TARGET_HOST}:${TARGET_PORT}"
+echo "Starting websocket proxy on ${LISTEN_HOST_PORT}"
+websockify "${LISTEN_HOST_PORT}" "${ADSB_HOST_PORT}"